Move openapi fs to internal pkg
