+++ internal/pkg/template/api/openapi/fs.go
@@ -1,10 +1,11 @@
 // nolint: gochecknoglobals
-package gen
+package openapi
 
 import (
 	"embed"
 )
 
+// Fs OpenAPI files are generated by "make generate-code"
 //go:embed openapi.json
 //go:embed openapi.yaml
 //go:embed openapi3.json
