+++ scripts/generate-templates-api.sh
@@ -30,7 +30,7 @@ generate() {
 finalize() {
     rm -rf ./internal/pkg/template/api/gen/http/cli
     rm -rf ./internal/pkg/template/api/gen/http/templates/client
-    mv ./internal/pkg/template/api/gen/http/openapi* ./api/templates/gen 2>/dev/null || true
+    mv ./internal/pkg/template/api/gen/http/openapi* ./internal/pkg/template/api/openapi 2>/dev/null || true
 }
 
 echo "Generating templates API ..."
